requirements.txt
torch>=1.7.0
numpy>=1.17.2,<2.0
scipy==1.6.0
hyperopt>=0.2.4
pandas>=1.0.5
tqdm>=4.48.2
scikit-learn>=0.23.2
pyyaml>=5.1.0
colorlog==4.7.2
colorama==0.4.4
tensorboard>=2.5.0
python-pptx>=0.6.21
